Add doCodechef.c source file
The doCodechef.c source file will be doing everything related to
'www.codechef.com' website. This will include parsing the web pages, submitting
requests etc.

As of now, the source is a simple program based on `libcurl`. It just outputs
the 'www.google.com' output page in text format.
This program basically tests the setup of the `libcurl` library.